% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk_code_get.R
\name{chunk_code_get}
\alias{chunk_code_get}
\title{Title}
\usage{
chunk_code_get(chunk_name = "chunk_code_get")
}
\arguments{
\item{chunk_name}{a character string with the name of the chunk of interest}
}
\value{
a vector of the code contained in the referenced chunk
}
\description{
Title
}
