% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk_to_dir.R
\name{chunk_to_dir}
\alias{chunk_to_dir}
\title{Title}
\usage{
chunk_to_dir(chunk_name, dir = "R/", extension = ".R")
}
\arguments{
\item{chunk_name}{}

\item{dir}{}

\item{extension}{}
}
\description{
Title
}
